Adding keyframei_t, player speed change, fix moving, adding debug gfx, ...
- Adding keyframei_t for keyframe interval control to have off sub-fields for single center tile.

Player position is now its center and hence weighted as required for the original behavior.

Since this also alows corner cutting for pacman, not waiting for a centered position to turn,
we can't rely on waiting for a center position for algorithms here.

Speed changes are implemented for different cases as layed out in <https://www.gamedeveloper.com/design/the-pac-man-dossier>.
It uses the keyframei_t::reset() method.
Both, ghosts and pacman are picking the !hint_slower for faster movement.
We have to think how to repaint a frame to sync the speed on get_fields_per_second_diff().

Debug graphics can all be switched on and include: target ray, grid and player bounds.

The SDL texture buggy offset has been compensated within 'texture_t::draw2()' .. don't ask :)