Fix ghost_t::set_next_dir(..) safeguard: Proceed if had collision (new) or if center position (old).
The next_field_frame_cntr count down has been fixed as well, but not used
since collision is a good enough criteria.

Note that the ghost_t directional change is originally intended to occur
only on center position of nodes with a decision (T or X section).

However, the acoord_t:step_impl() also doesn't handle a
D -> R transition after L -> R well since it brings back
the to be tested fwd_y_pos_i back to before the 'D' movement.
This is a weakness which might shall be addressed,
even though not hit with above constraints.